requirements.txt
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
matplotlib>=3.10.1
nltk>=3.9.1
numpy>=2.2.4
pandas>=2.2.3
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
sqlalchemy>=2.0.39
trafilatura>=2.0.0
yfinance>=0.2.55